requirements.txt
# DeltaMonitorBot - Ubuntu X11 only
# System dependency: tesseract-ocr (apt install tesseract-ocr)
mss>=9.0.0
Pillow>=10.0.0
pytesseract>=0.3.10
pyautogui>=0.9.54
